Changed decimal field to float field - 64 bits is plenty

+++ user_accounts/models.py
@@ -38,9 +38,9 @@ class User(AbstractUser):
     """Class to store user information in our database."""
 
     # Store LLM API credits.
-    # Decimal field to we can subtract fractions of a credit for small prompts.
+    # Float field to we can subtract fractions of a credit for small prompts.
     # Negative credits are permitted to allow for Tiktoken price estimation errors.
-    credits = models.DecimalField(max_digits=10, decimal_places=5, default=0)
-
+    credits = models.FloatField(default=0.0)
+    
     def __str__(self):
         return self.username
